split partnerIntegrations file into separate modules
