+++ src/tsconfig.strictNullChecks.json
@@ -95,8 +95,8 @@
     "./background/messenger/strict/api.ts",
     "./background/messenger/strict/registration.ts",
     "./background/partnerHandlers.ts",
-    "./background/partnerIntegrations.test.ts",
-    "./background/partnerIntegrations.ts",
+    "./background/partnerIntegrations/partnerIntegrations.test.ts",
+    "./background/partnerIntegrations/refreshPartnerToken.ts",
     "./background/performConfiguredRequest.test.ts",
     "./background/proxyUtils.ts",
     "./background/refreshRegistries.ts",
